Move mock data to subfolder (#8)
